feat: passport local 인증 추가, userRepository 추가, 리소스 접근시 권한체크 추가

+++ src/index.ts
@@ -1,9 +1,19 @@
 import express from 'express';
-import { xss } from 'express-xss-sanitizer';
+import session from 'express-session';
+import passport from 'passport';
+import { Strategy as LocalStrategy } from 'passport-local';
+import bcrypt from 'bcrypt';
+import dotenv from 'dotenv';
+import helmet from 'helmet';
 import path from 'node:path'
 import { fileURLToPath } from 'node:url';
 import { randomUUID } from 'node:crypto';
+import type { User } from './models/users.js';
 import { FilesystemPostRepository } from './repositories/postRepository.js';
+import { FilesystemUserRepository } from './repositories/userRepository.js';
+
+// 환경변수 불러오기
+dotenv.config();
 
 const app = express();
 const port = 3002;
@@ -21,9 +31,69 @@ app.set('views', path.join(__dirname, 'views'));
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
-// XSS 방지
-app.use(xss());
+// 보안 헤더
+app.use(helmet());
+// 세션 설정
+app.use(session({
+    secret: process.env.SESSION_SECRET_KEY || 'placeholder-secret-key',
+    resave: false,
+    saveUninitialized: false
+}));
 
+
+// 게시글 Repository 설정
+const postsDirectory = './src/data/posts';
+const postsRepository = new FilesystemPostRepository(postsDirectory);
+
+// 유저 Repository 설정
+const usersDirectory = './src/data/users';
+const usersRepository = new FilesystemUserRepository(usersDirectory);
+
+
+// passport 설정
+app.use(passport.initialize());
+app.use(passport.session());
+
+// passport local strategy 설정
+passport.use(new LocalStrategy({
+    usernameField: 'username',
+    passwordField: 'password'
+}, async (username, password, done) => { // verify 함수
+    try {
+
+        const user = await usersRepository.getUser(username);
+        if (!user) {
+            return done(null, false);
+        }
+
+        const isPasswordValid = await bcrypt.compare(password, user.hashed_password);
+        if (!isPasswordValid) {
+            return done(null, false);
+        }
+
+        return done(null, user);
+
+    } catch (err) {
+        return done(err);
+    }
+}));
+
+passport.serializeUser((user, done) => {
+    if ('username' in user && user.username !== undefined) {
+        done(null, user.username);
+    } else {
+        done(new Error('정상적인 User 객체가 아닙니다.'));
+    }
+});
+
+passport.deserializeUser(async (username: string, done) => {
+    try {
+        const user = await usersRepository.getUser(username);
+        done(null, user);
+    } catch (err) {
+        done(err);
+    }
+});
 
 app.get('/', (req, res) => {
     res.send("Hello!");
@@ -38,22 +108,55 @@ app.get('/params-example/:userSuppliedParameter', (req,res) => {
     res.render('index', {title: 'Params Example', body: paramsString});
 });
 
-// 게시글 Repository 설정
-const postsDirectory = './src/data/posts';
-const postsRepository = new FilesystemPostRepository(postsDirectory);
+// 로그인
+app.get('/login', (req, res) => {
+    res.render('login', {title: 'Login'});
+});
 
-// 게시글 CRUD
-app.get('/posts', async (req, res, next) => {
+app.post('/login', passport.authenticate('local', {successRedirect: '/posts', failureRedirect: '/login'}));
+
+// 로그인 강제 미들웨어
+function requireLogin(req: express.Request, res: express.Response, next: express.NextFunction) {
+    if (req.isAuthenticated()) {
+        return next();
+    }
+    res.redirect('/login');
+}
+
+// user 생성
+app.post('/users', async (req, res, next) => {
+    if (!req.body || !req.body.username || !req.body.password) {
+        return res.status(400).json({
+            error: '잘못된 요청입니다.',
+            message: '아이디와 비밀번호를 입력해주세요.'
+        });
+    }
+
     try {
-        const posts = await postsRepository.getAllPosts();
-        console.log(posts);
-        res.render('posts', {title: '게시글 목록', posts: posts});
+        await usersRepository.createUser(req.body.username, req.body.password);
+        res.redirect('/login');
     } catch (err) {
         next(err);
     }
 });
 
-app.post('/posts', async (req, res, next) => {
+// 게시글 CRUD
+app.get('/posts',
+    requireLogin,
+    async (req, res, next) => {
+        try {
+            const posts = await postsRepository.getAllPosts();
+            console.log(posts);
+            res.render('posts', {title: '게시글 목록', posts: posts});
+        } catch (err) {
+            next(err);
+        }
+    }
+);
+
+app.post('/posts',
+    requireLogin,
+    async (req, res, next) => {
     try {
         // 요청 데이터 검증
         if (!req.body || !req.body.title || !req.body.content) {
@@ -63,12 +166,20 @@ app.post('/posts', async (req, res, next) => {
             });
         }
 
+        if (!req.user || !('username' in req.user) || req.user.username === undefined || req.user.username === null) {
+            return res.status(400).json({
+                error: '잘못된 요청입니다.',
+                message: '로그인이 필요합니다.'
+            });
+        }
+
         const postId = String(randomUUID());
         const post = {
             id: postId,
             title: String(req.body.title),
             timestamp: new Date(),
-            content: String(req.body.content)
+            content: String(req.body.content),
+            createdBy: String(req.user.username)
         };
         
         await postsRepository.createPost(post);
@@ -79,7 +190,9 @@ app.post('/posts', async (req, res, next) => {
     }
 });
 
-app.get('/posts/new', async (req, res, next) => {
+app.get('/posts/new',
+    requireLogin,
+    async (req, res, next) => {
     try {
         res.render('new-post', {title: '새 게시글'});
     } catch (err) {
@@ -87,7 +200,9 @@ app.get('/posts/new', async (req, res, next) => {
     }
 });
 
-app.get('/posts/:postId', async (req,res) => {
+app.get('/posts/:postId',
+    requireLogin,
+    async (req,res) => {
     try {
         const postId = String(req.params.postId);
         const post = await postsRepository.getPost(postId);
@@ -104,8 +219,20 @@ app.patch('/posts/:postId', (req,res) => {
     throw new Error('아직 구현되지 않았습니다.');
 });
 
-app.delete('/posts/:postId', async (req, res) => {
+app.delete('/posts/:postId',
+    requireLogin,
+    async (req, res) => {
     try {
+        if (!req.user || !('username' in req.user) || req.user.username === undefined || req.user.username === null) {
+            throw new Error('로그인이 필요합니다.');
+        }
+        const post = await postsRepository.getPost(String(req.params.postId));
+        if (!post) {
+            throw new Error('게시글을 찾을 수 없습니다.');
+        }
+        if (post.createdBy !== req.user.username) {
+            throw new Error('자신의 게시글만 삭제할 수 있습니다.');
+        }
         await postsRepository.deletePost(String(req.params.postId));
         res.redirect(303, '/posts');
     } catch (err) {
